% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cern.R
\name{cern_record}
\alias{cern_record}
\title{Fetch a single CERN Open Data record by ID}
\usage{
cern_record(recid)
}
\arguments{
\item{recid}{Record ID (integer)}
}
\value{
tibble: one row with recid, title, type, experiment, date_published,
abstract, availability, publisher
}
\description{
Fetch a single CERN Open Data record by ID
}
